tree-wide: split open helpers into open_utils.h
