+++ src/lxc/lxccontainer.c
@@ -47,6 +47,7 @@
 #include "monitor.h"
 #include "namespace.h"
 #include "network.h"
+#include "open_utils.h"
 #include "parse.h"
 #include "process_utils.h"
 #include "start.h"
